Reorganize util folder
